Migrate propagate_block to external and add propagate_transaction
