+++ network/src/server/mod.rs
@@ -95,31 +95,3 @@ pub async fn process_transaction_internal(
 
     Ok(())
 }
-
-/// Broadcast block to connected peers
-pub async fn propagate_block(context: Arc<Context>, data: Vec<u8>, block_miner: SocketAddr) -> Result<(), SendError> {
-    debug!("Propagating a block to peers");
-
-    let peer_book = context.peer_book.read().await;
-    let local_address = *context.local_address.read().await;
-    let connections = context.connections.read().await;
-    let mut num_peers = 0;
-
-    for (socket, _) in &peer_book.get_connected() {
-        if *socket != block_miner && *socket != local_address {
-            if let Some(channel) = connections.get(socket) {
-                match channel.write(&Block::new(data.clone())).await {
-                    Ok(_) => num_peers += 1,
-                    Err(error) => warn!(
-                        "Failed to propagate block to peer {}. (error message: {})",
-                        channel.address, error
-                    ),
-                }
-            }
-        }
-    }
-
-    debug!("Block propagated to {} peers", num_peers);
-
-    Ok(())
-}
