+++ snarkos/miner.rs
@@ -16,7 +16,7 @@
 
 use snarkos_consensus::{ConsensusParameters, MemoryPool, MerkleTreeLedger, Miner};
 use snarkos_dpc::base_dpc::{instantiated::*, parameters::PublicParameters};
-use snarkos_network::server::{context::Context, propagate_block};
+use snarkos_network::{external::propagate_block, server::context::Context};
 use snarkos_objects::{AccountAddress, Block};
 
 use std::sync::Arc;
